Change the integration tests package name so it does not contain 'kotlinx' anymore
