+++ src/jvm/test-trace-debugger-integration/org/jetbrains/trace_debugger/integration/TraceDebuggerExamplesTraceDebuggerIntegrationTest.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.trace_debugger.integration
+package org.jetbrains.trace_debugger.integration
 
 import org.junit.Ignore
 import org.junit.Test
